+++ previous-versions/assess/pkg/models/shared/enhancedinvoicereportitem.go
@@ -3,11 +3,12 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type EnhancedInvoiceReportItem struct {
-	AmountDue *types.Decimal `json:"amountDue,omitempty"`
+	AmountDue *decimal.Big `decimal:"number" json:"amountDue,omitempty"`
 	// The currency data type in Codat is the [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217) currency code, e.g. _GBP_.
 	//
 	// ## Unknown currencies
@@ -91,10 +92,21 @@ type EnhancedInvoiceReportItem struct {
 	// - `Paid` - Invoice is paid in full. This includes if the invoice has been credited or overpaid (amountDue == 0).
 	// - `Void` - An invoice can become Void when it's deleted, refunded, written off, or cancelled. A voided invoice may still be PartiallyPaid, and so all outstanding amounts on voided invoices are removed from the accounts receivable account.
 	Status      *InvoiceStatus `json:"status,omitempty"`
-	TotalAmount *types.Decimal `json:"totalAmount,omitempty"`
+	TotalAmount *decimal.Big   `decimal:"number" json:"totalAmount,omitempty"`
 }
 
-func (o *EnhancedInvoiceReportItem) GetAmountDue() *types.Decimal {
+func (e EnhancedInvoiceReportItem) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(e, "", false)
+}
+
+func (e *EnhancedInvoiceReportItem) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &e, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *EnhancedInvoiceReportItem) GetAmountDue() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -178,7 +190,7 @@ func (o *EnhancedInvoiceReportItem) GetStatus() *InvoiceStatus {
 	return o.Status
 }
 
-func (o *EnhancedInvoiceReportItem) GetTotalAmount() *types.Decimal {
+func (o *EnhancedInvoiceReportItem) GetTotalAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}