+++ previous-versions/assess/RELEASES.md
@@ -371,3 +371,13 @@ Based on:
 - [go v0.25.0] previous-versions/assess
 ### Releases
 - [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.25.0 - previous-versions/assess
+
+## 2023-09-21 00:12:34
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.1] previous-versions/assess
+### Releases
+- [Go v0.25.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.25.1 - previous-versions/assess